Revert "Features/extract svg key"
