+++ package.json
@@ -32,11 +32,9 @@
     "@vitest/ui": "^1.2.0",
     "@vue/eslint-config-prettier": "^8.0.0",
     "@vue/test-utils": "^2.4.1",
-    "axios-mock-adapter": "^1.22.0",
     "eslint": "^8.49.0",
     "eslint-plugin-vue": "^9.17.0",
     "jsdom": "^22.1.0",
-    "msw": "^2.2.2",
     "prettier": "^3.0.3",
     "sass": "^1.69.5",
     "vite": "^4.5.1",
