+++ src/constants/apiVariables.js
@@ -1,0 +1,7 @@
+export const API_ENDPOINTS = {
+    login: '/login',
+    register: '/register',
+    createPost: '/create',
+    // other endpoints...
+  };
+
