+++ src/services/zoneService.js
@@ -1,14 +1,27 @@
-import { makeApiGetCall } from '@/api/api'
+import { makeApiGetCall } from '@/api' // Import the makeApiPostCall function
 import {  API_ENDPOINTS } from '@/constants/index.js'
 
+
+
+// const getZones = async () => {
+//     try {
+//       const response = await makeApiGetCall(API_ENDPOINTS.zone)
+//       return response.data.data
+//     } catch (error) {
+//       console.error('Error fetching all Zones:', error)
+//       throw error
+//     }
+//   }
 
   const getZones = async (level_id, parent_id) => {
     let params = new URLSearchParams();
   
+    // Add level_id to the params if it is not null
     if (level_id !== null && level_id !== undefined) {
       params.append('level_id', level_id);
     }
   
+    // Add parent_id to the params if it is not null
     if (parent_id !== null && parent_id !== undefined) {
       params.append('parent_id', parent_id);
     }
